Reorganize all files
* use linux-y folder names and hierarchy
* use linux-y include paths so that its easy to include other h's
