+++ student-distrib/Makefile
@@ -11,14 +11,14 @@ LDFLAGS += -nostdlib -static
 CC=gcc
 
 #If you have any .h files in another directory, add -I<dir> to this line
-CPPFLAGS +=-nostdinc -g
+CPPFLAGS +=-nostdinc -g -I include
 
 # This generates the list of source files
 SRC =  $(wildcard *.S) $(wildcard *.c) $(wildcard **/*.S) $(wildcard **/*.c)
 
 # This generates the list of .o files. The order matters, boot.o must be first
-OBJS  = boot.o
-OBJS += $(filter-out boot.o,$(patsubst %.S,%.o,$(filter %.S,$(SRC))))
+OBJS  = boot/boot.o
+OBJS += $(filter-out boot/boot.o,$(patsubst %.S,%.o,$(filter %.S,$(SRC))))
 OBJS += $(patsubst %.c,%.o,$(filter %.c,$(SRC)))
