+++ student-distrib/boot/x86_desc.S
@@ -2,7 +2,7 @@
 # vim:ts=4 noexpandtab
 
 #define ASM     1
-#include "x86_desc.h"
+#include <boot/x86_desc.h>
 
 .text
