+++ student-distrib/drivers/i8259.c
@@ -1,7 +1,7 @@
 /* i8259.c - Functions to interact with the 8259 interrupt controller
  */
 
-#include "i8259.h"
+#include <drivers/i8259.h>
 
 /* Interrupt masks to determine which interrupts
  * are enabled and disabled */
