+++ student-distrib/drivers/rtc.c
@@ -1,6 +1,6 @@
 /* rtc.c - Functions to interact with the real time clock
  */
-#include "rtc.h"
+#include <drivers/rtc.h>
 
 int32_t init_mode;
