+++ student-distrib/fs/files.c
@@ -1,4 +1,4 @@
-#include "files.h"
+#include <fs/files.h>
 
 boot_block b;
 uint32_t b_block_addrs;
