+++ student-distrib/include/boot/multiboot.h
@@ -14,7 +14,7 @@
 #ifndef ASM
 
 /* Types */
-#include "libs/types.h"
+#include <lib/types.h>
 
 /* The Multiboot header. */
 typedef struct multiboot_header
