+++ student-distrib/include/drivers/rtc.h
@@ -3,9 +3,9 @@
 #ifndef _RTC_H
 #define _RTC_H
 
-#include "../libs/lib.h"
-#include "i8259.h"
-#include "../libs/math.h"
+#include <lib/lib.h>
+#include <drivers/i8259.h>
+#include <lib/math.h>
 
 /* Constants for port selection */
 #define RTC_REG_NUM_PORT 0x70
