+++ student-distrib/kernel/user.S
@@ -1,5 +1,5 @@
 #define ASM 1
-#include "x86_desc.h"
+#include <boot/x86_desc.h>
 
 .global jmp_usr_exec
