+++ student-distrib/mm/paging.c
@@ -1,4 +1,4 @@
-#include "paging.h"
+#include <mm/paging.h>
 
 uint8_t * VIDEO_START;
